Move the cache functions from kubeadm to machine
