+++ pkg/minikube/bootstrapper/kubeadm/kubeadm.go
@@ -18,13 +18,10 @@ package kubeadm
 
 import (
 	"bytes"
-	"crypto"
 	"crypto/tls"
 	"fmt"
 	"net"
 	"net/http"
-	"os"
-	"path"
 	"strings"
 	"time"
 
@@ -32,7 +29,6 @@ import (
 	"github.com/docker/machine/libmachine"
 	"github.com/docker/machine/libmachine/state"
 	"github.com/golang/glog"
-	"github.com/jimmidyson/go-download"
 	"github.com/pkg/errors"
 	"golang.org/x/sync/errgroup"
 	"k8s.io/apimachinery/pkg/labels"
@@ -469,11 +465,11 @@ func (k *Bootstrapper) UpdateCluster(cfg config.KubernetesConfig) error {
 	for _, bin := range constants.GetKubeadmCachedBinaries() {
 		bin := bin
 		g.Go(func() error {
-			path, err := CacheBinary(bin, cfg.KubernetesVersion)
+			path, err := machine.CacheBinary(bin, cfg.KubernetesVersion)
 			if err != nil {
 				return errors.Wrapf(err, "downloading %s", bin)
 			}
-			err = CopyBinary(k.c, bin, path)
+			err = machine.CopyBinary(k.c, bin, path)
 			if err != nil {
 				return errors.Wrapf(err, "copying %s", bin)
 			}
@@ -579,51 +575,4 @@ func generateConfig(k8s config.KubernetesConfig, r cruntime.Manager) (string, er
 	}
 
 	return b.String(), nil
-}
-
-// CacheBinary will cache a binary on the host
-func CacheBinary(binary, version string) (string, error) {
-	targetDir := constants.MakeMiniPath("cache", version)
-	targetFilepath := path.Join(targetDir, binary)
-
-	url := constants.GetKubernetesReleaseURL(binary, version)
-
-	_, err := os.Stat(targetFilepath)
-	// If it exists, do no verification and continue
-	if err == nil {
-		glog.Infof("Not caching binary, using %s", url)
-		return targetFilepath, nil
-	}
-	if !os.IsNotExist(err) {
-		return "", errors.Wrapf(err, "stat %s version %s at %s", binary, version, targetDir)
-	}
-
-	if err = os.MkdirAll(targetDir, 0777); err != nil {
-		return "", errors.Wrapf(err, "mkdir %s", targetDir)
-	}
-
-	options := download.FileOptions{
-		Mkdirs: download.MkdirAll,
-	}
-
-	options.Checksum = constants.GetKubernetesReleaseURLSHA1(binary, version)
-	options.ChecksumHash = crypto.SHA1
-
-	console.OutStyle("file-download", "Downloading %s %s", binary, version)
-	if err := download.ToFile(url, targetFilepath, options); err != nil {
-		return "", errors.Wrapf(err, "Error downloading %s %s", binary, version)
-	}
-	return targetFilepath, nil
-}
-
-// CopyBinary copies previously cached binaries into the path
-func CopyBinary(cr bootstrapper.CommandRunner, binary, path string) error {
-	f, err := assets.NewFileAsset(path, "/usr/bin", binary, "0641")
-	if err != nil {
-		return errors.Wrap(err, "new file asset")
-	}
-	if err := cr.Copy(f); err != nil {
-		return errors.Wrapf(err, "copy")
-	}
-	return nil
 }
